renamed struct.* to value.*
