+++ src/app/lexer.l
@@ -4,7 +4,7 @@
 #include "parser.tab.h"
 
 #include "deimain.h"
-#include "struct.h"
+#include "value.h"
 #include "ast.h"
 #include "symbols.h"
