+++ src/app/parser.y
@@ -5,7 +5,7 @@
 #include "parser.tab.h"
 
 #include "deimain.h"
-#include "struct.h"
+#include "value.h"
 #include "ast.h"
 #include "symbols.h"
 #include "eval.h"
